+++ src/decorators/Arg.ts
@@ -1,24 +1,28 @@
 import { MetadataStorage } from "../metadata/metadata-storage";
-import { ReturnTypeFunc, ClassType, TypeOptions } from "../types/decorators";
+import { ReturnTypeFunc, ClassType, BasicOptions } from "../types/decorators";
 import { getParamInfo } from "../helpers/params";
-import { getDecoratorParams } from "../helpers/decorators";
+import { getTypeDecoratorParams } from "../helpers/decorators";
 
-export function Arg(name: string, typeOptions?: TypeOptions): ParameterDecorator;
+export function Arg(name: string, BasicOptions?: BasicOptions): ParameterDecorator;
 export function Arg(
   name: string,
   returnTypeFunc: ReturnTypeFunc,
-  typeOptions?: TypeOptions,
+  BasicOptions?: BasicOptions,
 ): ParameterDecorator;
 export function Arg(
   name: string,
-  returnTypeFuncOrOptions?: ReturnTypeFunc | TypeOptions,
-  maybeOptions?: TypeOptions,
+  returnTypeFuncOrOptions?: ReturnTypeFunc | BasicOptions,
+  maybeOptions?: BasicOptions,
 ): ParameterDecorator {
   return (prototype, propertyKey, parameterIndex) => {
-    const { options, returnTypeFunc } = getDecoratorParams(returnTypeFuncOrOptions, maybeOptions);
+    const { options, returnTypeFunc } = getTypeDecoratorParams(
+      returnTypeFuncOrOptions,
+      maybeOptions,
+    );
     MetadataStorage.registerHandlerParam({
       kind: "arg",
       name,
+      description: options.description,
       ...getParamInfo(prototype, propertyKey, parameterIndex, returnTypeFunc, options),
     });
   };