+++ src/decorators/GraphQLResolver.ts
@@ -1,6 +1,7 @@
 import { MetadataStorage } from "../metadata/metadata-storage";
+import { ClassType } from "../types/decorators";
 
-export function GraphQLResolver(typeFunc: () => any): ClassDecorator {
+export function GraphQLResolver(typeFunc: () => ClassType): ClassDecorator {
   return target => {
     MetadataStorage.registerResolver({
       target,