+++ src/helpers/decorators.ts
@@ -1,9 +1,13 @@
-import { TypeOptions, ReturnTypeFunc } from "../types/decorators";
+import { TypeOptions, ReturnTypeFunc, DescriptionOptions } from "../types/decorators";
 
-export function getDecoratorParams(
-  returnTypeFuncOrOptions: ReturnTypeFunc | TypeOptions | undefined,
-  maybeOptions: TypeOptions | undefined,
-) {
+export interface TypeDecoratorParams<T> {
+  options: Partial<T>;
+  returnTypeFunc?: ReturnTypeFunc;
+}
+export function getTypeDecoratorParams<T extends TypeOptions>(
+  returnTypeFuncOrOptions: ReturnTypeFunc | T | undefined,
+  maybeOptions: T | undefined,
+): TypeDecoratorParams<T> {
   if (typeof returnTypeFuncOrOptions === "function") {
     return {
       returnTypeFunc: returnTypeFuncOrOptions,
@@ -12,6 +16,22 @@ export function getDecoratorParams(
   } else {
     return {
       options: returnTypeFuncOrOptions || {},
+    };
+  }
+}
+
+export function getNameDecoratorParams(
+  nameOrOptions: string | DescriptionOptions | undefined,
+  maybeOptions: DescriptionOptions | undefined,
+) {
+  if (typeof nameOrOptions === "string") {
+    return {
+      name: nameOrOptions,
+      options: maybeOptions || {},
+    };
+  } else {
+    return {
+      options: nameOrOptions || {},
     };
   }
 }